Remove somewhat false-confidence building code coverage implementation.
We do need code coverage, but the comment within the code that I'm removing
basically states the obstacle we're now running into - again.

(See the link to https://stackoverflow.com/questions/61989746/how-do-i-use-zprofile-to-generate-code-coverage-for-a-nostd-kernel)

The introduction of this was arguably a bit premature given the relative
infancy of the Rust code coverage tooling in the ecosystem and some of the
inefficient practices in the way that this is conveying coverage today
(without the help of LLVM).  Hope is on the way (see attached), but for now
let's just count on the fact that Rust's compiler is pretty good.

Ref: https://blog.rust-lang.org/inside-rust/2020/11/12/source-based-code-coverage.html